feat: change token management for users to be outside the scope of the lib.
BREAKING CHANGES:
Use Services.AddHttpClient("AuthClient", (sp, httpClient) => ...).AddUserAccessTokenHandler(); or any other kinds of token management, and then use "AuthClient" to inform lib which client it should use.
CgScriptOptions.Site is no longer used. Set the BaseAddress of the httpClient in the above config to decide which site to call.